backend/requirements.txt
# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# AI能力
dashscope==1.14.1
openai==1.12.0

# 爬虫
scrapy==2.11.0
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.1

# Excel处理
openpyxl==3.1.2
xlsxwriter==3.1.9

# 阿里云SDK
oss2==2.18.4
alibabacloud-tea-openapi==0.3.9

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.26.0
apscheduler==3.10.4
python-multipart==0.0.6

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0
